Added env variables to configure the wireguard socket directory
The wireguard socket directory defaults to '/var/run/wireguard'. On some
platforms the /var/run directory is not available. With this commit, the
wireguard socket directoy can be configured with the environment
variable WG_SOCKET_DIR. If the environment variable is not set or empty,
the default '/var/run/wireguard' is used.